+++ packages/fighting-design/checkbox-group/src/checkbox-group.vue
@@ -1,27 +1,54 @@
 <script lang="ts" setup name="FCheckboxGroup">
-  import { nextTick, provide, reactive, toRefs } from 'vue'
-  import { Props, Emits, checkboxGroupCtxKey } from './checkbox-group'
-  import type { CheckboxLabelType } from './interface'
+  import { provide, reactive, toRefs, computed } from 'vue'
+  import { Props, Emits, checkboxGroupPropsKey } from './checkbox-group'
+  import { sizeChange } from '../../_utils'
+  import type {
+    CheckboxGroupLabelType as a,
+    ChangeEventInterface as b
+  } from './interface'
+  import type { ComputedRef, CSSProperties } from 'vue'
+  import type { ClassListInterface as c } from '../../_interface'
 
   const prop = defineProps(Props)
   const emit = defineEmits(Emits)
 
-  const changeEvent = async (val: CheckboxLabelType): Promise<void> => {
+  const changeEvent: b = (val: a): void => {
     emit('update:modelValue', val)
-    await nextTick()
     emit('change', val)
   }
 
-  const checkboxGroup = reactive({
+  const checkboxGroupProps = reactive({
     ...toRefs(prop),
     changeEvent
+  } as const)
+
+  provide(checkboxGroupPropsKey, checkboxGroupProps)
+
+  const styleList: ComputedRef<CSSProperties> = computed((): CSSProperties => {
+    const { columnGap, rowGap } = prop
+
+    return {
+      columnGap: sizeChange(columnGap),
+      rowGap: sizeChange(rowGap)
+    } as const
   })
 
-  provide(checkboxGroupCtxKey, checkboxGroup)
+  const classList: ComputedRef<c> = computed((): c => {
+    const { border, vertical, size } = prop
+
+    return [
+      'f-checkbox-group',
+      {
+        'f-checkbox-group-border': border,
+        'f-checkbox-group-vertical': vertical,
+        [`f-checkbox-group-${size}`]: size && border
+      }
+    ] as const
+  })
 </script>
 
 <template>
-  <div class="f-checkbox-group">
+  <div role="group" :class="classList" :style="styleList">
     <slot />
   </div>
 </template>