+++ packages/fighting-design/checkbox-group/src/interface.d.ts
@@ -1,3 +1,7 @@
 export type CheckboxGroupLabelType = string | number | boolean
 
-export type CheckboxGroupSizeType = 'large' | 'middle' | 'small'
+export type CheckboxGroupSizeType = 'large' | 'middle' | 'small' | 'mini'
+
+export interface ChangeEventInterface {
+  (val: CheckboxGroupLabelType): void
+}