+++ packages/fighting-design/dialog/src/dialog.vue
@@ -41,6 +41,7 @@
     >
       <div
         v-show="visible"
+        role="dialog"
         :class="[
           'f-dialog-mask',
           {