+++ packages/fighting-design/image/src/image.vue
@@ -76,7 +76,7 @@
 <template>
   <div
     v-if="isSuccess"
-    rule="img"
+    role="img"
     :class="['f-image', { 'f-image-block': block }]"
     :style="styleList"
   >