+++ packages/fighting-design/rate/src/interface.d.ts
@@ -1,0 +1,7 @@
+export interface OnMouseoverInterface {
+  (index: number): void
+}
+
+export interface OnHandleClickInterface {
+  (index: number): void
+}