+++ start/src/App.vue
@@ -1,11 +1,11 @@
 <script lang="ts" setup>
   import { ref } from 'vue'
 
-  const value1 = ref(3)
+  const value3 = ref(2)
 </script>
 
 <template>
-  <f-rate v-model="value1" double-clear />
+  <f-rate v-model="value3" text-show icon="f-icon-atm-fill" />
 </template>
 
 <style lang="scss" scoped></style>